+++ sdk/base-macros/src/generate.rs
@@ -548,6 +548,7 @@ fn push_default_with_entries(opts: &mut Opts) {
     push_path_entry(opts, &format!("{CORE_TYPES_INTERFACE}/recipient"), "::miden::Recipient");
     push_path_entry(opts, &format!("{CORE_TYPES_INTERFACE}/note-idx"), "::miden::NoteIdx");
     push_path_entry(opts, &format!("{CORE_TYPES_INTERFACE}/nonce"), "::miden::Nonce");
+    push_path_entry(opts, &format!("{CORE_TYPES_INTERFACE}/block-number"), "::miden::BlockNumber");
 }
 
 fn push_path_entry(opts: &mut Opts, key: &str, value: &str) {